requirements.txt
asgiref==3.7.2
boto3==1.28.26
botocore==1.31.26
Django==4.2.4
django-bootstrap-submenu==0.1.7
django-ckeditor==6.7.0
django-jalali==6.0.1
django-jalali-date==1.0.2
django-js-asset==2.1.0
django-mptt==0.14.0
django-storages==1.13.2
fontawesome==5.10.1.post1
fontawesomefree==6.4.2
jdatetime==4.1.1
jmespath==1.0.1
Pillow==10.0.0
psycopg2-binary==2.9.7
python-dateutil==2.8.2
s3transfer==0.6.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==1.26.16

requests~=2.31.0